backend/requirements.txt
google-adk>=1.18.0
google-genai>=1.47.0
google-auth>=2.49.0
google-auth-oauthlib>=1.3.0
google-api-python-client>=2.192.0
fastapi>=0.115.0
uvicorn[standard]>=0.39.0
websockets>=15.0.0
python-dotenv>=1.1.0
pillow>=11.0.0
httpx>=0.27.0
